Move the new functions into a separate storage module
