+++ sway-core/src/ir_generation/convert.rs
@@ -6,10 +6,8 @@ use crate::{
 
 use super::types::{create_enum_aggregate, create_tuple_aggregate};
 
-use fuel_crypto::Hasher;
-use fuel_types::Bytes32;
 use sway_ir::{Aggregate, Constant, Context, MetadataIndex, Type, Value};
-use sway_types::{span::Span, state::StateIndex};
+use sway_types::span::Span;
 
 use std::sync::Arc;
 
@@ -141,35 +139,4 @@ fn convert_resolved_type(
         TypeInfo::ErrorRecovery => reject_type!("Error recovery"),
         TypeInfo::Storage { .. } => reject_type!("Storage"),
     })
-}
-
-use uint::construct_uint;
-
-#[allow(
-// These two warnings are generated by the `construct_uint!()` macro below.
-    clippy::assign_op_pattern,
-    clippy::ptr_offset_with_cast
-)]
-pub(super) fn add_to_b256(x: fuel_types::Bytes32, y: u64) -> fuel_types::Bytes32 {
-    construct_uint! {
-        struct U256(4);
-    }
-    let x = U256::from(*x);
-    let y = U256::from(y);
-    let res: [u8; 32] = (x + y).into();
-    fuel_types::Bytes32::from(res)
-}
-
-pub(super) fn get_storage_key<T>(ix: &StateIndex, indices: &[T]) -> Bytes32
-where
-    T: std::fmt::Display,
-{
-    Hasher::hash(indices.iter().fold(
-        format!(
-            "{}{}",
-            sway_utils::constants::STORAGE_DOMAIN_SEPARATOR,
-            ix.to_usize()
-        ),
-        |acc, i| format!("{}_{}", acc, i),
-    ))
 }
